Reorder với custom sales

+++ LibMagestore/src/main/java/com/magestore/app/pos/model/sales/PosOrderCartItem.java
@@ -59,7 +59,7 @@ public class PosOrderCartItem extends PosAbstractModel implements OrderCartItem 
 
     @Override
     public int getQty() {
-        return Integer.parseInt(qty);
+        return (int) Float.parseFloat(qty);
     }
 
     @Override
